+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add signing key's corresponding public key and algorithm to the `Signature` proto message
 - Fixes to Couchbase Plugin
 - Add "signature rotation errors" field to `UpdateApiSignature` and `UpdateApplicationSignature` proto messages (move existing `Signature` into `result` oneof field)
+- Added `prefix` and `delimeter` support to S3 Plugin 
 
 ## v1.15.1